fixed draw sort order

+++ Platformer v0.3/Bitmask.cpp
@@ -9,14 +9,14 @@ void Bitmask::SetMask(Bitmask& other)
 	bits = other.GetMask();
 }
 
-uint32_t Bitmask::GetMask() const
+uint64_t Bitmask::GetMask() const
 {
 	return bits;
 }
 
 bool Bitmask::GetBit(int pos) const
 {
-	return (bits & (1 << pos)) != 0;
+	return (bits & (static_cast<unsigned long long>(1) << pos)) != 0;
 }
 
 void Bitmask::SetBit(int pos, bool on)
@@ -33,12 +33,12 @@ void Bitmask::SetBit(int pos, bool on)
 
 void Bitmask::SetBit(int pos)
 {
-	bits = bits | 1 << pos;
+	bits = bits | static_cast<unsigned long long>(1) << pos;
 }
 
 void Bitmask::ClearBit(int pos)
 {
-	bits = bits & ~(1 << pos);
+	bits = bits & ~(static_cast<unsigned long long>(1) << pos);
 }
 
 void Bitmask::Clear()
